views: add update_job_status, refactor views

+++ ssherlock/ssherlock_server/views.py
@@ -1,23 +1,12 @@
 """All Django views for the SSHerlock server application."""
 
-# pylint: disable=import-error
-from django.http import Http404
-from django.http import JsonResponse
+from django.http import Http404, JsonResponse, HttpResponse
 from django.views.decorators.http import require_http_methods
-from django.shortcuts import get_object_or_404
-from django.shortcuts import redirect
-from django.shortcuts import render
-
-from .forms import BastionHostForm
-from .forms import CredentialForm
-from .forms import JobForm
-from .forms import LlmApiForm
-from .forms import TargetHostForm
-from .models import BastionHost
-from .models import Credential
-from .models import Job
-from .models import LlmApi
-from .models import TargetHost
+from django.shortcuts import get_object_or_404, redirect, render
+from .forms import BastionHostForm, CredentialForm, JobForm, LlmApiForm, TargetHostForm
+from .models import BastionHost, Credential, Job, LlmApi, TargetHost
+from .utils import check_private_key
+import json
 
 
 def landing(request):
@@ -37,17 +26,13 @@ MODEL_FORM_MAP = {
 
 def handle_object(request, model_type, uuid=None):
     """Handle creating or editing any object except jobs."""
-    # Get the model, form, and template based on the model_type parameter
-    try:
-        model, form = MODEL_FORM_MAP.get(model_type)
-    except TypeError as exc:
-        raise Http404("Model type not found.") from exc
+    model_form_tuple = MODEL_FORM_MAP.get(model_type)
+    if not model_form_tuple:
+        raise Http404("Model type not found.")
 
-    instance = None
-    if uuid:
-        instance = get_object_or_404(model, pk=uuid)
-
-    form = form(request.POST or None, instance=instance)
+    model, form_class = model_form_tuple
+    instance = get_object_or_404(model, pk=uuid) if uuid else None
+    form = form_class(request.POST or None, instance=instance)
 
     if request.method == "POST" and form.is_valid():
         form.save()
@@ -58,47 +43,37 @@ def handle_object(request, model_type, uuid=None):
         "object_name": model_type.capitalize(),
         "uuid": uuid,
     }
-    if uuid:
-        template_name = "ssherlock_server/objects/edit_object.html"
-    else:
-        template_name = "ssherlock_server/objects/add_object.html"
+    template_name = (
+        "ssherlock_server/objects/edit_object.html"
+        if uuid
+        else "ssherlock_server/objects/add_object.html"
+    )
     return render(request, template_name, context)
 
 
-def delete_object(request, model_type, uuid):  # pylint: disable=unused-argument
+def delete_object(request, model_type, uuid):
     """Delete the given object."""
-    # Get the model, form, and template based on the model_type parameter
-    try:
-        model, _ = MODEL_FORM_MAP.get(model_type)
-    except TypeError as exc:
-        raise Http404("Model type not found.") from exc
+    model_form_tuple = MODEL_FORM_MAP.get(model_type)
+    if not model_form_tuple:
+        raise Http404("Model type not found.")
 
+    model, _ = model_form_tuple
     instance = get_object_or_404(model, pk=uuid)
-
     instance.delete()
     return redirect(f"/{model_type}_list")
 
 
 def create_job(request):
-    """Handle creating jobs.
-
-    When the create job form is submitted, a new job is created for every target host.
-    """
+    """Handle creating jobs. When the create job form is submitted, a new job is created for every target host."""
     form = JobForm(request.POST)
-
     if request.method == "POST" and form.is_valid():
-
         cleaned_data = form.cleaned_data
-        # Get list of target hosts and also remove the list of target hosts from the cleaned data.
         target_hosts = cleaned_data.pop("target_hosts", [])
 
         for host in target_hosts:
-            # Copy all the attributes from cleaned data into the job (excluding target hosts).
             job = Job(**cleaned_data)
-            # Save the job object to create a primary key.
             job.save()
-            # Add the single host to each job that's created.
-            job.target_hosts.add(host.id)  # pylint: disable=no-member
+            job.target_hosts.add(host.id)
             job.save()
 
         return redirect("/job_list")
@@ -107,8 +82,7 @@ def create_job(request):
         "form": form,
         "object_name": "Job",
     }
-    template_name = "ssherlock_server/objects/add_object.html"
-    return render(request, template_name, context)
+    return render(request, "ssherlock_server/objects/add_object.html", context)
 
 
 def home(request):
@@ -118,82 +92,50 @@ def home(request):
 
 def bastion_host_list(request):
     """List the bastion hosts."""
-    output = BastionHost.objects.all()  # pylint: disable=no-member
-    column_headers = ["Hostname"]
-    object_fields = ["hostname"]
-    object_name = "Bastion Host"
-    context = {
-        "output": output,
-        "column_headers": column_headers,
-        "object_name": object_name,
-        "object_fields": object_fields,
-    }
-    return render(request, "ssherlock_server/objects/object_list.html", context)
+    return render_object_list(
+        request, BastionHost, ["Hostname", "Port"], ["hostname", "port"], "Bastion Host"
+    )
 
 
 def credential_list(request):
     """List the credentials."""
-    output = Credential.objects.all()  # pylint: disable=no-member
-    column_headers = ["Name", "Username", "Password"]
-    object_fields = ["credential_name", "username", "password"]
-    object_name = "Credential"
-    context = {
-        "output": output,
-        "column_headers": column_headers,
-        "object_name": object_name,
-        "object_fields": object_fields,
-    }
-    return render(request, "ssherlock_server/objects/object_list.html", context)
+    return render_object_list(
+        request,
+        Credential,
+        ["Name", "Username", "Password"],
+        ["credential_name", "username", "password"],
+        "Credential",
+    )
 
 
 def llm_api_list(request):
     """List the LLM APIs."""
-    output = LlmApi.objects.all()  # pylint: disable=no-member
-    column_headers = ["Base URL", "API Key"]
-    object_fields = ["base_url", "api_key"]
-    object_name = "LLM API"
-    context = {
-        "output": output,
-        "column_headers": column_headers,
-        "object_name": object_name,
-        "object_fields": object_fields,
-    }
-    return render(request, "ssherlock_server/objects/object_list.html", context)
+    return render_object_list(
+        request, LlmApi, ["Base URL", "API Key"], ["base_url", "api_key"], "LLM API"
+    )
 
 
 def job_list(request):
     """List the jobs."""
-    output = Job.objects.all()  # pylint: disable=no-member
-
-    column_headers = [
-        "Status",
-        "LLM API",
-        "Target Hosts",
-        "Target Host Credentials",
-    ]
-    object_fields = [
-        "status",
-        "llm_api",
-        "target_hosts",
-        "credentials_for_target_hosts",
-    ]
-
-    object_name = "Job"
-    context = {
-        "output": output,
-        "column_headers": column_headers,
-        "object_name": object_name,
-        "object_fields": object_fields,
-    }
-    return render(request, "ssherlock_server/objects/object_list.html", context)
+    return render_object_list(
+        request,
+        Job,
+        ["Status", "LLM API", "Target Hosts", "Target Host Credentials"],
+        ["status", "llm_api", "target_hosts", "credentials_for_target_hosts"],
+        "Job",
+    )
 
 
 def target_host_list(request):
     """List the target hosts."""
-    output = TargetHost.objects.all()  # pylint: disable=no-member
-    column_headers = ["Hostname"]
-    object_fields = ["hostname"]
-    object_name = "Target Host"
+    return render_object_list(
+        request, TargetHost, ["Hostname", "Port"], ["hostname", "port"], "Target Host"
+    )
+
+
+def render_object_list(request, model, column_headers, object_fields, object_name):
+    """Helper function to render object lists."""
+    output = model.objects.all()
     context = {
         "output": output,
         "column_headers": column_headers,
@@ -204,29 +146,17 @@ def target_host_list(request):
 
 
 @require_http_methods(["GET"])
-def request_job(request):  # pylint: disable=unused-argument
-    """
-    Provide a job for runners to process.
-
-    This is the API endpoint used by runners to retrieve a job.
-    """
+def request_job(request):
+    """Provide a job for runners to process. This is the API endpoint used by runners to retrieve a job."""
     try:
-        # Only return data if the correct private key is in the header.
-        private_key = request.headers.get('Private-Key')
-        if not private_key:
-            return JsonResponse({"message": "Private key not provided."}, status=404)
-        elif private_key != "myprivatekey":
-            return JsonResponse({"message": "Private key incorrect."}, status=404)
+        key_check_response = check_private_key(request)
+        if key_check_response:
+            return key_check_response
 
-        # Query the oldest pending job
-        job = (
-            Job.objects.filter(status="PENDING").order_by("created_at").first()
-        )  # pylint: disable=no-member
-
-        if job is None:
+        job = Job.objects.filter(status="PENDING").order_by("created_at").first()
+        if not job:
             return JsonResponse({"message": "No pending jobs found."}, status=404)
 
-        # Use getattr here since some fields may not have values.
         job_data = {
             "id": str(job.id),
             "llm_api_baseurl": getattr(job.llm_api, "base_url", None),
@@ -239,10 +169,8 @@ def request_job(request):  # pylint: disable=unused-argument
             "credentials_for_bastion_host_password": getattr(
                 job.credentials_for_bastion_host, "password", None
             ),
-            "target_host_hostname": getattr(
-                job.target_hosts.all()[0], "hostname", None
-            ),
-            "target_host_port": getattr(job.target_hosts.all()[0], "port", None),
+            "target_host_hostname": getattr(job.target_hosts.first(), "hostname", None),
+            "target_host_port": getattr(job.target_hosts.first(), "port", None),
             "credentials_for_target_hosts_username": getattr(
                 job.credentials_for_target_hosts, "username", None
             ),
@@ -253,6 +181,29 @@ def request_job(request):  # pylint: disable=unused-argument
         }
 
         return JsonResponse(job_data, status=200)
+
+    except Exception as e:
+        return JsonResponse({"message": str(e)}, status=500)
+
+
+@require_http_methods(["POST"])
+def update_job_status(request, job_id):
+    """Update the status of a job. This is the API endpoint used by runners to update the status of a job."""
+    try:
+        key_check_response = check_private_key(request)
+        if key_check_response:
+            return key_check_response
+
+        data = json.loads(request.body)
+        new_status = data.get("status")
+        if not new_status:
+            return JsonResponse({"message": "Status not provided."}, status=400)
+
+        job = get_object_or_404(Job, pk=job_id)
+        job.status = new_status
+        job.save()
+
+        return HttpResponse(status=200)
 
     except Exception as e:
         return JsonResponse({"message": str(e)}, status=500)
